refactor: config builder file organization
