+++ client/src/pages/Configurations/ConfigBuild/CodeSets/AddConditionCodeSetsDrawer.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { AddConditionCodeSetsDrawer } from './AddConditionCodeSets';
+import { AddConditionCodeSetsDrawer } from './AddConditionCodeSetsDrawer';
 import userEvent from '@testing-library/user-event';
 
 // Mocks for child components and hooks
