+++ client/src/pages/Configurations/ConfigBuild/CodeSets/Drawer.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { Drawer } from '.';
+import { Drawer } from './Drawer';
 import userEvent from '@testing-library/user-event';
 import { ClassAttributes, InputHTMLAttributes } from 'react';
 import { JSX } from 'react/jsx-runtime';
