+++ client/src/pages/Configurations/ConfigBuild/Lock/ConfigLockBanner.tsx
